pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "illume"
version = "1.0.0"
description = "Illuminating Unified MLLM with Dual Visual Tokenization and Diffusion Refinement"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    # install torch first
    "torch==2.6.0", "torchvision",
    "deepspeed==0.16.9",
    "transformers==4.44.2", "tokenizers",
    "attrdict", "diffusers==0.31.0",
    "sentencepiece==0.1.99",
    "shortuuid",
    "accelerate", "peft", "bitsandbytes",
    "requests", "httpx==0.23.3", "uvicorn", "fastapi",
    "einops==0.6.1", "einops-exts==0.0.4",
    "timm==0.9.12", "tensorboardX", "tensorboard", "tiktoken",
    "loguru",
    "gradio==3.50.2", "gradio-client==0.6.1",
    "pydantic==2.11.5",
    "pydantic_core==2.33.2",
    "markdown2[all]", "numpy", "scikit-learn==1.2.2",
    "einx==0.3.0", "Pillow==9.0.1",
    "tenacity", "sqlitedict", "evaluate", "sacrebleu", "hf_transfer",
    "scikit-image", "torch_fidelity", "imagesize"
]

[project.optional-dependencies]
npu = [
    "pyyaml", "numpy==1.23.5", "decorator", "scipy", "attrs", "psutil",
    "torch_npu==2.4.0",
]
train = ["ninja", "wandb"]
build = ["build", "twine"]

[tool.setuptools.packages.find]
exclude = ["assets*", "configs*", "evaluations*", "dist*", "scripts*"]

[tool.wheel]
exclude = ["assets*", "configs*", "evaluations*", "dist*", "scripts*"]
